requirements.txt
beautifulsoup4==4.4.1
cmsplugin-polls==0.1.0
cmsplugin-zinnia==0.8
dj-database-url==0.4.1
Django==1.9.5
django-app-namespace-template-loader==0.3.1
django-appconf==1.0.2
django-blog-zinnia==0.16
django-classy-tags==0.7.2
django-cms==3.2.3
django-cms-wunderground==1.0.0
django-compressor==2.0
django-contrib-comments==1.7.0
django-debug-toolbar==1.4
django-formtools==1.0
django-haystack==2.4.1
django-mptt==0.8.3
django-reversion==1.10.1
django-sekizai==0.9.0
Django-Select2==4.3.2
django-tagging==0.4.1
django-treebeard==4.0
django-xmlrpc==0.1.5
djangocms-admin-style==1.1.0
djangocms-column==1.6.0
djangocms-file==1.0
djangocms-googlemap==0.4.0
djangocms-inherit==0.1.1
djangocms-installer==0.8.7
djangocms-link==1.7.2
djangocms-picture==0.2.0
djangocms-style==1.7.0
djangocms-teaser==0.2.0
djangocms-text-ckeditor==2.9.1
djangocms-video==1.0.0
gunicorn==19.4.5
html5lib==0.9999999
mots-vides==2015.5.11
mysqlclient==1.3.7
Pillow==3.2.0
pyparsing==2.1.1
python-memcached==1.57
pytz==2016.3
rcssmin==1.0.6
rjsmin==1.0.12
six==1.10.0
sqlparse==0.1.19
tzlocal==1.2.2
Whoosh==2.7.4
zinnia-theme-bootstrap==0.4
zinnia-wysiwyg-wymeditor==1.1
